requirements.txt
anthropic>=0.25.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
schedule>=1.2.0
flask>=3.0.0
